ai/GRU_models/pip_requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
etils==0.9.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.2
Flask==2.2.2
Flask-Cors==3.0.10
flask-restx==1.0.3
flatbuffers==22.11.23
flit_core==3.8.0
fonttools==4.38.0
gast==0.4.0
gensim==4.2.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.46.3
h5py==3.6.0
htmlmin==0.1.12
huggingface-hub==0.11.1
idna==3.4
ImageHash==4.3.1
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
JPype1==1.4.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==5.1.0
jupyter-events==0.5.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
jupyterthemes==0.20.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
lesscpy==0.15.1
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mecab-python==0.996-ko-0.9.2
mistune==2.0.4
multimethod==1.9
munkres==1.1.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandas-profiling==3.5.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.0
ply==3.11
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.1
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
setuptools==65.5.1
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metadata==1.12.0
tensorflow-metal==0.6.0
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
transformers==4.25.1
typeguard==2.13.3
typing_extensions==4.4.0
unicodedata2==15.0.0
urllib3==1.26.13
visions==0.7.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wheel==0.38.4
widgetsnbextension==4.0.4
wrapt==1.14.1
zipp==3.11.0
